requirements.txt
flask_sqlalchemy
flask_wtf
flask_bcrypt

you need to pip install these to run this project
